+++ roadgraph_builder/io/lidar/__init__.py
@@ -9,11 +9,14 @@ For centerline-offset lane ribbons without point clouds, see
 """
 
 from roadgraph_builder.io.lidar.fusion import attach_lidar_points_metadata, fuse_lane_boundaries_from_points
+from roadgraph_builder.io.lidar.las import LASHeader, read_las_header
 from roadgraph_builder.io.lidar.loader import load_lidar_placeholder, load_points_xy_csv
 
 __all__ = [
+    "LASHeader",
     "attach_lidar_points_metadata",
     "fuse_lane_boundaries_from_points",
     "load_lidar_placeholder",
     "load_points_xy_csv",
+    "read_las_header",
 ]